backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.2.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Transcription & Audio Processing
openai-whisper==20231117
ffmpeg-python==0.2.0
pydub==0.25.1
speechrecognition==3.10.0

# File handling
python-magic==0.4.27
aiofiles==23.2.1
python-docx==1.1.0
PyPDF2==3.0.1
rq==1.15.1
redis==5.0.1

# API Documentation
swagger-ui-bundle==0.0.9

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0